requirements.txt
Flask>=2.2
pymongo>=4.0
flask-pymongo>=2.3
bcrypt>=4.0
PyJWT>=2.8
flask-cors>=4.0
openai>=1.0.0
google-generativeai>=0.3.0
pydub>=0.25.1
python-dotenv>=1.0.0
certifi>=2023.0.0
